Move output key computation into InnerTemporaryStore method
